chore: version packages (#244)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.17.2

### Patch Changes

-   7e01577: fix(react): remove `<Await/>`

## @suspensive/react-await@0.0.1

### Patch Changes

-   7e01577: fix(react): remove `<Await/>`

## @suspensive/react-query@1.17.2

### Patch Changes

-   Updated dependencies [7e01577]
    -   @suspensive/react@1.17.2

Co-authored-by: Jonghyeon Ko <manudeli.ko@gmail.com>
Co-authored-by: Chanhyuk Park <pgg6713@gmail.com>
Co-authored-by: Minsoo Kim <zlemzlem5656@naver.com>