chore: version packages (#368)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.20.2

### Patch Changes

-   e37fbcc: chore(deps-dev): Bump @types/react from 18.2.34 to 18.2.37

## @suspensive/react-await@0.0.8

### Patch Changes

-   e37fbcc: chore(deps-dev): Bump @types/react from 18.2.34 to 18.2.37

## @suspensive/react-query@1.20.2

### Patch Changes

-   e37fbcc: chore(deps-dev): Bump @types/react from 18.2.34 to 18.2.37
-   Updated dependencies [e37fbcc]
    -   @suspensive/react@1.20.2

Co-authored-by: Jonghyeon Ko <manudeli.ko@gmail.com>
Co-authored-by: Chanhyuk Park <pgg6713@gmail.com>
Co-authored-by: Minsoo Kim <zlemzlem5656@naver.com>