chore: version packages (#393)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react-image@0.0.1

### Patch Changes

-   cd225ab: feat(react-image): initialize @suspensive/react-image