chore: version packages (#402)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.20.5

### Patch Changes

-   b5888b2: fix(react): make wrap as stable and deprecate all withXXXs
-   bce7c1d: fix(react): deprecate AsyncBoundary

## @suspensive/react-query@1.20.5

### Patch Changes

-   bce7c1d: fix(react): deprecate AsyncBoundary
-   Updated dependencies [b5888b2]
-   Updated dependencies [bce7c1d]
    -   @suspensive/react@1.20.5