chore: version packages (#458)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.20.7

### Patch Changes

-   b69d651: fix(tsup): add supported browsers

## @suspensive/react-await@0.0.11

### Patch Changes

-   b69d651: fix(tsup): add supported browsers

## @suspensive/react-image@0.0.3

### Patch Changes

-   b69d651: fix(tsup): add supported browsers

## @suspensive/react-query@1.20.7

### Patch Changes

-   b69d651: fix(tsup): add supported browsers
-   Updated dependencies [b69d651]
    -   @suspensive/react@1.20.7

## @suspensive/visualization@0.0.9

### Patch Changes

-   Updated dependencies [b69d651]
    -   @suspensive/react@1.20.7
    -   @suspensive/react-query@1.20.7
    -   @suspensive/react-await@0.0.11
    -   @suspensive/react-image@0.0.3