chore: version packages (#500)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.20.9

### Patch Changes

-   685054a: ci(eslint): add import/no-cycle to prevent memory leak

## @suspensive/react-query@1.20.9

### Patch Changes

-   Updated dependencies [685054a]
    -   @suspensive/react@1.20.9

## @suspensive/visualization@0.0.11

### Patch Changes

-   Updated dependencies [685054a]
    -   @suspensive/react@1.20.9
    -   @suspensive/react-await@0.0.12
    -   @suspensive/react-query@1.20.9