chore: version packages (#546)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.22.2

### Patch Changes

- f56c40e: fix(react): add OmitKeyOf to remove PropsWithoutChildren,
PropsWithoutDevMode

## @suspensive/react-query@1.22.2

### Patch Changes

-   Updated dependencies [f56c40e]
    -   @suspensive/react@1.22.2

## @suspensive/visualization@0.0.16

### Patch Changes

-   Updated dependencies [f56c40e]
    -   @suspensive/react@1.22.2
    -   @suspensive/react-await@0.0.14
    -   @suspensive/react-query@1.22.2