chore: version packages (#552)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.23.0

### Minor Changes

-   e776d52: feat(react): add svg logo for `<DevMode/>`

## @suspensive/react-query@1.23.0

### Patch Changes

-   Updated dependencies [e776d52]
    -   @suspensive/react@1.23.0

## @suspensive/visualization@0.0.17

### Patch Changes

-   Updated dependencies [e776d52]
    -   @suspensive/react@1.23.0
    -   @suspensive/react-await@0.0.14
    -   @suspensive/react-query@1.23.0