chore: version packages (#556)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react-image@0.0.7

### Patch Changes

-   05a3534: refactor(react-image): rename Image -> SuspenseImage

## @suspensive/visualization@0.0.18

### Patch Changes

-   Updated dependencies [05a3534]
    -   @suspensive/react-image@0.0.7