chore: version packages (#581)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.23.1

### Patch Changes

-   c526ad3: fix(react): assert message correctly, test code simply

## @suspensive/react-query@1.23.1

### Patch Changes

-   Updated dependencies [c526ad3]
    -   @suspensive/react@1.23.1

## @suspensive/visualization@0.0.19

### Patch Changes

-   Updated dependencies [c526ad3]
    -   @suspensive/react@1.23.1
    -   @suspensive/react-await@0.0.14
    -   @suspensive/react-query@1.23.1