chore: version packages (#597)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.24.0

### Minor Changes

-   89c3283: feat(react): add fallback prop of Delay

## @suspensive/react-query@1.24.0

### Patch Changes

-   Updated dependencies [89c3283]
    -   @suspensive/react@1.24.0

## @suspensive/visualization@0.0.21

### Patch Changes

-   Updated dependencies [89c3283]
    -   @suspensive/react@1.24.0
    -   @suspensive/react-await@0.0.15
    -   @suspensive/react-query@1.24.0