chore: version packages (#603)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.25.0

### Minor Changes

-   caed129: feat(react): add `shouldCatch` prop of ErrorBoundary

## @suspensive/react-query@1.25.0

### Patch Changes

-   Updated dependencies [caed129]
    -   @suspensive/react@1.25.0

## @suspensive/visualization@0.0.22

### Patch Changes

-   Updated dependencies [caed129]
    -   @suspensive/react@1.25.0
    -   @suspensive/react-await@0.0.15
    -   @suspensive/react-query@1.25.0