chore: version packages (#609)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react-image@0.0.9

### Patch Changes

-   794e07f: fix(react-image): throw Error on error during image loading

## @suspensive/visualization@0.0.23

### Patch Changes

-   Updated dependencies [794e07f]
    -   @suspensive/react-image@0.0.9