chore: version packages (#637)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.25.1

### Patch Changes

- 37f5465: feat(react): add ErrorBoundary.Consumer,
ErrorBoundaryGroup.Consumer

## @suspensive/react-query@1.25.1

### Patch Changes

-   Updated dependencies [37f5465]
    -   @suspensive/react@1.25.1

## @suspensive/visualization@0.0.24

### Patch Changes

-   Updated dependencies [37f5465]
    -   @suspensive/react@1.25.1
    -   @suspensive/react-await@0.0.15
    -   @suspensive/react-query@1.25.1