chore: version packages (#638)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.25.2

### Patch Changes

-   a6d279e: fix(\*): add use-sync-external-store

## @suspensive/react-await@0.0.16

### Patch Changes

-   a6d279e: fix(\*): add use-sync-external-store

## @suspensive/react-image@0.0.10

### Patch Changes

-   a6d279e: fix(\*): add use-sync-external-store

## @suspensive/react-query@1.25.2

### Patch Changes

-   Updated dependencies [a6d279e]
    -   @suspensive/react@1.25.2

## @suspensive/visualization@0.0.25

### Patch Changes

-   Updated dependencies [a6d279e]
    -   @suspensive/react-await@0.0.16
    -   @suspensive/react-image@0.0.10
    -   @suspensive/react@1.25.2
    -   @suspensive/react-query@1.25.2