chore: version packages (#657)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @suspensive/react@1.25.3

### Patch Changes

- af96ce0: fix(react): add getServerSnapshot in for syncDevMode in SSR
environment

## @suspensive/react-query@1.25.3

### Patch Changes

-   Updated dependencies [af96ce0]
    -   @suspensive/react@1.25.3

## @suspensive/next-streaming-react-query@0.1.1

### Patch Changes

-   Updated dependencies [af96ce0]
    -   @suspensive/react@1.25.3

## @suspensive/visualization@0.0.26

### Patch Changes

-   Updated dependencies [af96ce0]
    -   @suspensive/react@1.25.3
    -   @suspensive/react-await@0.0.16
    -   @suspensive/react-query@1.25.3