feat(utility-types): add isNullable utility type (#731)
# Overview

This is a minor suggestion! 🙏

How about managing types that have `null` as a union type with an
`isNullable` utility type?

These types are meaningful and could be reused enough in the future. 

However, this task may be subject to personal preferences. **The opinion
of the maintainer is important...!!💪**

In the case of `vitest`, it also includes `undefined` as an `isNullable`
type, but we don't seem to need it.

```tsx
// vitest
type Nullable<T> = T | null | undefined;
```

## PR Checklist

- [x] I did below actions if need

1. I read the [Contributing
Guide](https://github.com/suspensive/react/blob/main/CONTRIBUTING.md)
2. I added documents and tests.

---------

Co-authored-by: Jonghyeon Ko <manudeli.ko@gmail.com>